% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectADPRN.R
\docType{class}
\name{TrenaProjectADPRN-class}
\alias{TrenaProjectADPRN-class}
\alias{.TrenaProjectADPRN}
\alias{TrenaProjectADPRN}
\title{TrenaProjectADPRN-class}
\usage{
TrenaProjectADPRN(quiet = TRUE)
}
\value{
An object of the TrenaProjectADPRN class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
